requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0          # async postgres driver
aiosqlite==0.20.0        # async sqlite (for tests)
alembic==1.13.1          # database migrations

# Settings
pydantic-settings==2.5.2
pydantic[email]==2.9.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1             # distributed caching & rate limiting

# HTTP client (external API integration)
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Code quality
ruff==0.6.9
